+++ src/dev.config.ts
@@ -14,7 +14,7 @@ export const DEV_STARTING_COINS = 200;
 
 export const USE_FIXED_PLANTS = true;
 const sharedDebugConfig = {
-  hue: -4,
+  hue: 255,
   petalCount: 6,
   petalEffect: 'shimmer' as PetalEffect,
   plantPhase: 3 as PlantPhase,
@@ -68,7 +68,7 @@ export const DEBUG_PLANTS = [
 
 export const USE_FIXED_SEEDS = true;
 const sharedSeedConfig = {
-  hue: -4,
+  hue: -2,
   petalCount: 6,
   plantPhase: 3 as PlantPhase,
 }